Move SyncError to pontoon.sync.exceptions
