+++ pontoon/sync/__init__.py
@@ -1,6 +1,2 @@
-class SyncError(RuntimeError):
-    """Error class for errors relating to the project sync process."""
-
-
 # A separator used in Translation keys as proposed by Translate Toolkit.
 KEY_SEPARATOR = "\x04"
