+++ pontoon/sync/exceptions.py
@@ -1,2 +1,6 @@
 class ParseError(RuntimeError):
     """Exception to raise when parsing fails."""
+
+
+class SyncError(RuntimeError):
+    """Error class for general errors relating to the project sync process."""
