+++ pontoon/sync/formats/ftl.py
@@ -6,7 +6,7 @@ import logging
 
 from fluent.syntax import ast, FluentParser, FluentSerializer
 
-from pontoon.sync import SyncError
+from pontoon.sync.exceptions import SyncError
 from pontoon.sync.formats.base import ParsedResource
 from pontoon.sync.utils import create_parent_directory
 from pontoon.sync.vcs.models import VCSTranslation
