+++ pontoon/sync/formats/json_extensions.py
@@ -16,8 +16,7 @@ from collections import OrderedDict
 from jsonschema import validate
 from jsonschema.exceptions import ValidationError
 
-from pontoon.sync import SyncError
-from pontoon.sync.exceptions import ParseError
+from pontoon.sync.exceptions import ParseError, SyncError
 from pontoon.sync.formats.base import ParsedResource
 from pontoon.sync.utils import create_parent_directory
 from pontoon.sync.vcs.models import VCSTranslation
